// src/shared/api/models/getApiProductsId200InventoryTransactionItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiProductsId200InventoryTransactionItemTransactionType } from './getApiProductsId200InventoryTransactionItemTransactionType';

export type GetApiProductsId200InventoryTransactionItem = {
    transactionType: GetApiProductsId200InventoryTransactionItemTransactionType;
    quantity: number;
    unitCost: number;
    unitPrice: number;
    createdAt: string;
    customerId: string;
    customerName: string;
    employeeId: string;
    updatedBy: string;
    employeeName: string;
    inStock: number;
};